Kela Clinic/requirements.txt
appdirs
asgiref
astroid
atomicwrites
attrs
autopep8
beautifulsoup4
category-encoders
certifi
chardet
colorama
coverage
cycler
distlib
dj-database-url
django-bootstrap-staticfiles
django-chartjs
django-contrib-comments
django-crispy-forms
django-jenkins
django-jquery
django-load
django-mysql
django-pandas
django-staticfiles
filelock
future
gunicorn
idna
importlib-metadata
iniconfig
isort
joblib
jsons
kiwisolver
lazy-object-proxy
#matplotlib==3.4.2
mccabe
#mysql==0.0.2
mysql-connector
mysql-connector-python
#mysqlclient==2.0.1
numpy
opencv-contrib-python
packaging
pandas
patsy
pep8
Pillow
pluggy
protobuf
urllib3
values
virtualenv
whitenoise
wrapt
zipp
selenium
#opencv-contrib-python-headless
#traitlets
django-debug-toolbar
